+++ AzureDevOpsBackup/Class/DisplayHelp.cs
@@ -19,7 +19,10 @@ namespace AzureDevOpsBackup.Class
             Console.WriteLine();
             Console.WriteLine("Parameter List:");
             Console.WriteLine("  Mandatory:");
-            Console.WriteLine("\t--token:             Token to access the API in Azure DevOps");
+            Console.WriteLine("\t--token <token>:     Token to access the API in Azure DevOps (raw token data)");
+            Console.WriteLine("\t--token token.bin:   Will use data to access the API in Azure DevOps from encrypted token.bin file");
+            Console.WriteLine("\t--tokenfile <token>: Save a token to access the API in Azure DevOps to an encrypted token.bin file");
+            Console.WriteLine("\t\t\t     (use before using the '--token token.bin' argument!)");
             Console.WriteLine("\t--org:               Name of the organization in Azure DevOps");
             Console.WriteLine("\t--backup:            Folder where to store the backup(s) - folder with timestamp will be created");
             Console.WriteLine("\t--server:            IP address or DNS name of the SMTP server");
@@ -52,6 +55,8 @@ namespace AzureDevOpsBackup.Class
             Console.WriteLine($"\t{Globals._currentExeFileName} --token XXX... --org OrgName --backup C:\\Backup --server smtp.domain.local");
             Console.WriteLine("\t--port 25 --from from@domain.local --to reports@domain.local --unzip --simpelreport\n");
             Console.WriteLine($"\t{Globals._currentExeFileName} --token XXX... --org OrgName --backup C:\\Backup --server smtp.domain.local");
+            Console.WriteLine("\t--port 25 --from from@domain.local --to reports@domain.local --unzip --simpelreport --priority high\n");
+            Console.WriteLine($"\t{Globals._currentExeFileName} --token token.bin --org OrgName --backup C:\\Backup --server smtp.domain.local");
             Console.WriteLine("\t--port 25 --from from@domain.local --to reports@domain.local --unzip --simpelreport --priority high");
             Console.WriteLine();
             Console.WriteLine("Output:");